requirements.txt
asgiref==3.5.2
autopep8==1.7.0
backports.zoneinfo==0.2.1
Django==4.1.1
django-ckeditor==6.5.1
django-filter==22.1
django-js-asset==2.0.0
django-rest-framework==0.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
Pillow==9.2.0
psycopg2-binary==2.9.3
pycodestyle==2.9.1
PyJWT==2.5.0
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
